requirements.txt
fastapi
uvicorn
gunicorn
pyodbc
pandas
azure-identity
